created cli module and moved select as submodule
